Implementation of the database in the GUI
Reading functions have been created

the “Analytics division” page has been linked to the database

+++ Blosoft_DatabaseManager/DatabaseManager.hpp
@@ -4,7 +4,7 @@
 #include <string>
 
 struct DataEntry {
-    int timestamp;
+    unsigned long timestamp;
     int dataQuantity;
 };
